Add "iso cluster" for DPU provisioning
Add support to deploy a cluster of type iso, which takes an arbitrary
iso file as an additional argument, and provisions a single host with
this ISO.

The ISO can be a local file, or an http url.

There are a number of assumptions made about the iso and BMC such that
this currently will only work for the DPU use-case, however in the
future this can be made more general to be hardware agnostic.

Example Config:

clusters:
- name: "iso-cluster"
  api_vip: "0.0.0.0"
  ingress_vip: "0.0.0.0"
  network_api_port: "eno12409"
  kind: "iso"
  install_iso: "http://myhost.redhat.com/rhel.iso"
  masters:
  - name: "acc-239"
    node: "foobar"
    type: "physical"
    bmc: "idrac.hostname.redhat.com"
    bmc_user: "root"
    bmc_password: "calvin"
    ip: "192.168.3.24"
    mac: "10:23:00:03:b5:42"

"network_api_port" should be an interface on the provisioning host
connected to the node in question.

"mac" should be the mac address of the port on the single host connected
to the "network_api_port"

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>