Remove workaround to enable ACC connectivity
We should not need to manually run cfg_acc_apf_x2.py. If the user is
installing for the first time, we will expect to be on the default
package, meaning this is not needed.

If the IMC has been reconfigured to use a custom package, we should
expect the vsp has installed a method to ensure connectivity is up on
ACC reboot.

Signed-off-by: Salvatore Daniele <sdaniele@redhat.com>